Update CSS
CSS updated to include a phone media query as well as tablet size. 
Increased with of page for phone to almost 100vw instead of 50vw (Following trend of other mobile sites. 
Had to change initial height of divs to accommodate for overflow in inner-author div. 
Adjusted font-size for new queries.